refactor: composants type bannière de simulation
