+++ site/source/components/Simulation/SimulationPréremplieBanner.tsx
@@ -2,13 +2,13 @@ import { Trans, useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
 import { styled } from 'styled-components'
 
+import SimulationBanner from '@/components/Simulation/Banner'
 import { PopoverWithTrigger } from '@/design-system'
 import { Link } from '@/design-system/typography/link'
 import { companySituationSelector } from '@/store/selectors/simulationSelectors'
 
-import Banner from './Banner'
-import AnswerList from './conversation/AnswerList'
-import WrongSimulateurWarning from './WrongSimulateurWarning'
+import AnswerList from '../conversation/AnswerList'
+import WrongSimulateurWarning from '../WrongSimulateurWarning'
 
 export default function SimulationPréremplieBanner() {
 	const existingCompany = !!useSelector(companySituationSelector)[
@@ -22,7 +22,7 @@ export default function SimulationPréremplieBanner() {
 	}
 
 	return (
-		<Banner icon="✏">
+		<SimulationBanner icon="✏">
 			<Trans i18nKey="simulationPréremplieBanner.info">
 				Ce simulateur a été prérempli avec la situation de votre entreprise.
 			</Trans>{' '}
@@ -47,7 +47,7 @@ export default function SimulationPréremplieBanner() {
 			<WrongSimulateurWarningContainer>
 				<WrongSimulateurWarning />
 			</WrongSimulateurWarningContainer>
-		</Banner>
+		</SimulationBanner>
 	)
 }
