+++ site/source/pages/simulateurs/artiste-auteur/ArtisteAuteur.tsx
@@ -3,12 +3,12 @@ import { useLocation } from 'react-router-dom'
 
 import { Condition } from '@/components/EngineValue/Condition'
 import PeriodSwitch from '@/components/PeriodSwitch'
-import { SelectSimulationYear } from '@/components/SelectSimulationYear'
 import SimulateurWarning from '@/components/SimulateurWarning'
 import Simulation, {
 	SimulationGoal,
 	SimulationGoals,
 } from '@/components/Simulation'
+import { YearSelectionBanner } from '@/components/Simulation/YearSelectionBanner'
 import { DistributionBranch } from '@/components/simulationExplanation/DistributionDesCotisations'
 import { InstitutionsPartenairesArtisteAuteur } from '@/components/simulationExplanation/InstitutionsPartenaires'
 import { useEngine } from '@/components/utils/EngineContext'
@@ -27,7 +27,7 @@ export default function ArtisteAuteur() {
 			<Simulation
 				results={<InstitutionsPartenairesArtisteAuteur />}
 				explanations={<CotisationsResult />}
-				afterQuestionsSlot={<SelectSimulationYear />}
+				afterQuestionsSlot={<YearSelectionBanner />}
 			>
 				<SimulateurWarning
 					simulateur="artiste-auteur"
