+++ site/source/pages/simulateurs/indépendant/EntrepriseIndividuelle.tsx
@@ -1,6 +1,6 @@
-import { SelectSimulationYear } from '@/components/SelectSimulationYear'
 import SimulateurWarning from '@/components/SimulateurWarning'
 import Simulation from '@/components/Simulation'
+import { YearSelectionBanner } from '@/components/Simulation/YearSelectionBanner'
 import IndépendantExplanation from '@/components/simulationExplanation/IndépendantExplanation'
 import { IndépendantSimulationGoals } from '@/pages/simulateurs/indépendant/Goals'
 
@@ -8,7 +8,7 @@ export const EntrepriseIndividuelle = () => (
 	<>
 		<Simulation
 			explanations={<IndépendantExplanation />}
-			afterQuestionsSlot={<SelectSimulationYear />}
+			afterQuestionsSlot={<YearSelectionBanner />}
 		>
 			<SimulateurWarning simulateur="entreprise-individuelle" />
 			<IndépendantSimulationGoals legend="Vos revenus d'entreprise individuelle" />
