+++ site/source/pages/simulateurs/reduction-generale/RéductionGénérale.tsx
@@ -4,9 +4,9 @@ import { Trans, useTranslation } from 'react-i18next'
 import PeriodSwitch from '@/components/PeriodSwitch'
 import EffectifSwitch from '@/components/RéductionDeCotisations/EffectifSwitch'
 import RégularisationSwitch from '@/components/RéductionDeCotisations/RégularisationSwitch'
-import { SelectSimulationYear } from '@/components/SelectSimulationYear'
 import SimulateurWarning from '@/components/SimulateurWarning'
 import Simulation from '@/components/Simulation'
+import { YearSelectionBanner } from '@/components/Simulation/YearSelectionBanner'
 import { Strong } from '@/design-system/typography'
 import { Body } from '@/design-system/typography/paragraphs'
 import { RégularisationMethod } from '@/utils/réductionDeCotisations'
@@ -45,7 +45,7 @@ export default function RéductionGénéraleSimulation() {
 
 	return (
 		<>
-			<Simulation afterQuestionsSlot={<SelectSimulationYear />}>
+			<Simulation afterQuestionsSlot={<YearSelectionBanner />}>
 				<SimulateurWarning
 					simulateur="réduction-générale"
 					informationsComplémentaires={
